# salesforce-deployer/src/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: your.email@example.com
  description: |-
    API for extracting, managing, and deploying Salesforce metadata schemas.
    This API enables seamless management of Salesforce configurations across environments.
  license:
    name: MIT
  title: Salesforce Schema Deployer API
  version: 0.1.0
servers:
- description: Development server
  url: http://localhost:8000/api/v1
- description: Mock server for testing
  url: http://localhost:8001/api/v1
...
paths:
  /auth/login:
    post:
      description: "Authenticates a user with their Salesforce username, password,\
        \ and optional security token to obtain an access token for API access."
      operationId: authenticateWithCredentials
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          description: Authentication successful
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed
      summary: Authenticate with Salesforce credentials
      tags:
      - authentication
  /auth/oauth:
    post:
      description: "Authenticates using the OAuth 2.0 protocol, supporting both web\
        \ server flow (with auth code) and username-password flow for headless integrations."
      operationId: authenticateWithOAuth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          description: Authentication successful
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Authentication failed
      summary: Authenticate using OAuth 2.0
      tags:
      - authentication
  /auth/refresh:
    post:
      description: Obtains a new access token using a valid refresh token when the
        current token has expired or is about to expire.
      operationId: refreshToken
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          description: Token refreshed successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid or expired refresh token
      security:
      - BearerAuth: []
      summary: Refresh authentication token
      tags:
      - authentication
...
  /deployments:
    post:
      description: Initiates a deployment process to apply schema changes to a target
        Salesforce organization with configurable validation and rollback options.
      operationId: deploySchema
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentRequest'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
          description: Deployment started
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Deploy schema to a Salesforce organization
      tags:
      - deployment
  /deployments/{deploymentId}:
    get:
      description: "Retrieves the current status, progress, and results of an ongoing\
        \ or completed deployment operation."
      operationId: getDeploymentStatus
      parameters:
      - explode: false
        in: path
        name: deploymentId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentStatus'
          description: Deployment status found
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Deployment not found
      security:
      - BearerAuth: []
      summary: Get deployment status by ID
      tags:
      - deployment
  /deployments/{deploymentId}/cancel:
    post:
      description: "Attempts to cancel an in-progress deployment operation, rolling\
        \ back any changes if possible based on the deployment phase."
      operationId: cancelDeployment
      parameters:
      - explode: false
        in: path
        name: deploymentId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentStatus'
          description: Deployment cancelled successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Deployment cannot be cancelled
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Deployment not found
      security:
      - BearerAuth: []
      summary: Cancel a deployment
      tags:
      - deployment
  /metadata/objects:
    get:
      description: Retrieves a list of all available objects from either a stored
        schema or directly from a connected Salesforce organization.
      operationId: getObjects
      parameters:
      - description: "ID of the schema to get objects from. If not provided, objects\
          \ from the connected Salesforce org will be returned."
        explode: true
        in: query
        name: schemaId
        required: false
        schema:
          format: uuid
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ObjectMetadata'
                type: array
          description: List of objects retrieved successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Get a list of objects
      tags:
      - metadata
  /metadata/objects/{objectName}:
    get:
      description: "Fetches comprehensive metadata for a specific object, including\
        \ its fields, relationships, and configuration properties."
      operationId: getObjectByName
      parameters:
      - explode: false
        in: path
        name: objectName
        required: true
        schema:
          type: string
        style: simple
      - description: "ID of the schema to get the object from. If not provided, the\
          \ object from the connected Salesforce org will be returned."
        explode: true
        in: query
        name: schemaId
        required: false
        schema:
          format: uuid
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectMetadata'
          description: Object details retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Object not found
      security:
      - BearerAuth: []
      summary: Get object details by name
      tags:
      - metadata
  /metadata/objects/{objectName}/fields:
    get:
      description: "Returns a list of all fields defined for a specific object, with\
        \ their data types, validation rules, and relationships."
      operationId: getObjectFields
      parameters:
      - explode: false
        in: path
        name: objectName
        required: true
        schema:
          type: string
        style: simple
      - description: "ID of the schema to get fields from. If not provided, fields\
          \ from the connected Salesforce org will be returned."
        explode: true
        in: query
        name: schemaId
        required: false
        schema:
          format: uuid
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FieldMetadata'
                type: array
          description: Fields retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Object not found
      security:
      - BearerAuth: []
      summary: Get fields for an object
      tags:
      - metadata
...
components:
  schemas:
    LoginRequest:
      example:
        securityToken: ABCDEFGHIJKLMNOP
        password: myPassword123
        username: user@example.com
        instanceUrl: https://login.salesforce.com
      properties:
        username:
          description: Salesforce username
          example: user@example.com
          title: username
          type: string
        password:
          description: Salesforce password
          example: myPassword123
          format: password
          title: password
          type: string
        securityToken:
          description: Salesforce security token
          example: ABCDEFGHIJKLMNOP
          title: securityToken
          type: string
        instanceUrl:
          description: Salesforce instance URL
          example: https://login.salesforce.com
          title: instanceUrl
          type: string
      required:
      - password
      - username
      title: LoginRequest
      type: object
    OAuthRequest:
      example:
        redirectUri: https://localhost:8000/callback
        password: myPassword123
        clientId: 3MVG9e2mBbZnmcolVXiRm8YNot7VI5W_oiqJuBDdDUNF9s45MfoiV2h_nI2paMT1QA_yw3UxQncI7Goz8L1A7
        authCode: aPrxJZTWTs65VeNEEhYht92nXyz
        clientSecret: "123456789012345678"
        username: user@example.com
      properties:
        clientId:
          description: OAuth client ID
          example: 3MVG9e2mBbZnmcolVXiRm8YNot7VI5W_oiqJuBDdDUNF9s45MfoiV2h_nI2paMT1QA_yw3UxQncI7Goz8L1A7
          title: clientId
          type: string
        clientSecret:
          description: OAuth client secret
          example: "123456789012345678"
          title: clientSecret
          type: string
        redirectUri:
          description: OAuth redirect URI
          example: https://localhost:8000/callback
          title: redirectUri
          type: string
        authCode:
          description: OAuth authorization code (required if not using username/password
            flow)
          example: aPrxJZTWTs65VeNEEhYht92nXyz
          title: authCode
          type: string
        username:
          description: Salesforce username (required if using username/password flow)
          example: user@example.com
          title: username
          type: string
        password:
          description: Salesforce password (required if using username/password flow)
          example: myPassword123
          format: password
          title: password
          type: string
      required:
      - clientId
      - clientSecret
      title: OAuthRequest
      type: object
    AuthenticationResponse:
      example:
        expiresIn: 3600
        accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        instanceUrl: https://na1.salesforce.com
      properties:
        accessToken:
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          title: accessToken
          type: string
        refreshToken:
          description: JWT refresh token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          title: refreshToken
          type: string
        instanceUrl:
          description: Salesforce instance URL
          example: https://na1.salesforce.com
          title: instanceUrl
          type: string
        expiresIn:
          description: Token expiration time in seconds
          example: 3600
          title: expiresIn
          type: integer
      required:
      - accessToken
      - instanceUrl
      title: AuthenticationResponse
      type: object
    ErrorResponse:
      example:
        details:
        - field: username
          message: Field is required
        - field: username
          message: Field is required
        error: INVALID_SESSION_ID
        message: Session expired or invalid
      properties:
        error:
          description: Error code
          example: INVALID_SESSION_ID
          title: error
          type: string
        message:
          description: Error message
          example: Session expired or invalid
          title: message
          type: string
        details:
          description: Detailed error information
          items:
            $ref: '#/components/schemas/ErrorResponse_details_inner'
          title: details
          type: array
      required:
      - error
      - message
      title: ErrorResponse
      type: object
...
    DeploymentRequest:
      example:
        schema:
          createdAt: 2025-04-21T14:30:00Z
          components:
            workflowRules:
              key:
                criteria: Account.Status__c = 'Active'
                name: Account_Status_Update
                active: true
                description: Updates Account status based on criteria
                actions:
                - field: Status__c
                  name: UpdateAccountStatus
                  type: FieldUpdate
                  value: Processed
                - field: Status__c
                  name: UpdateAccountStatus
                  type: FieldUpdate
                  value: Processed
                object: Account
            flows:
              key:
                name: Lead_Process
                description: Automates lead qualification
                label: Lead Qualification Process
                processType: AutoLaunchedFlow
                triggerType: ""
                eventType: ""
                status: Active
            objects:
              key:
                custom: false
                name: Account
                description: Standard Account object
                nameField: Name
                label: Account
                fields:
                  key:
                    relationshipName: ""
                    defaultValue: Default value
                    unique: false
                    name: CustomField__c
                    length: 255
                    externalId: false
                    description: A custom text field
                    formula: ""
                    label: Custom Field
                    type: Text
                    required: false
                    referenceTo: ""
                sharingModel: ReadWrite
            dashboards:
              key:
                components:
                - reportName: SalesReport
                  name: SalesChart
                  type: Chart
                - reportName: SalesReport
                  name: SalesChart
                  type: Chart
                name: Sales_Dashboard
                description: Key sales metrics dashboard
                label: Sales Performance
                folderName: Sales Dashboards
          apiVersion: "56.0"
          sourceOrgId: 123e4567-e89b-12d3-a456-426614174000
          name: Production Schema April 2025
          description: Current production schema as of April 2025
          id: 123e4567-e89b-12d3-a456-426614174000
          updatedAt: 2025-04-21T14:30:00Z
        components:
        - Account.CustomField__c
        - Account.CustomField__c
        targetOrgId: 223e4567-e89b-12d3-a456-426614174000
        schemaId: 123e4567-e89b-12d3-a456-426614174000
        options:
          testLevel: NoTestRun
          ignoreWarnings: false
          runTests: false
          rollbackOnError: true
          autoUpdatePackage: false
          purgeOnDelete: false
          checkOnly: false
      properties:
        schemaId:
          description: ID of the schema to deploy
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: schemaId
          type: string
        schema:
          $ref: '#/components/schemas/Schema'
        targetOrgId:
          description: ID of the target organization
          example: 223e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: targetOrgId
          type: string
        components:
          description: "List of components to deploy. If not provided, all components\
            \ will be deployed."
          items:
            example: Account.CustomField__c
            type: string
          title: components
          type: array
        options:
          $ref: '#/components/schemas/DeploymentRequest_options'
      required:
      - targetOrgId
      title: DeploymentRequest
      type: object
    DeploymentResponse:
      example:
        validationResults:
          valid: true
          dependencyGraph:
            key:
            - dependencyGraph
            - dependencyGraph
          warnings:
          - severity: error
            componentType: field
            path: objects.Account.fields.CustomField__c
            componentName: Account.CustomField__c
            message: Field 'CustomField__c' references non-existent object 'NonExistentObject__c'
          - severity: error
            componentType: field
            path: objects.Account.fields.CustomField__c
            componentName: Account.CustomField__c
            message: Field 'CustomField__c' references non-existent object 'NonExistentObject__c'
          errors:
          - severity: error
            componentType: field
            path: objects.Account.fields.CustomField__c
            componentName: Account.CustomField__c
            message: Field 'CustomField__c' references non-existent object 'NonExistentObject__c'
          - severity: error
            componentType: field
            path: objects.Account.fields.CustomField__c
            componentName: Account.CustomField__c
            message: Field 'CustomField__c' references non-existent object 'NonExistentObject__c'
        deploymentId: 123e4567-e89b-12d3-a456-426614174000
        message: Deployment queued successfully
        status: queued
      properties:
        deploymentId:
          description: Deployment ID
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: deploymentId
          type: string
        status:
          description: Deployment status
          enum:
          - queued
          - in_progress
          - completed
          - failed
          - cancelled
          example: queued
          title: status
          type: string
        message:
          description: Additional information
          example: Deployment queued successfully
          title: message
          type: string
        validationResults:
          $ref: '#/components/schemas/SchemaValidationResponse'
      required:
      - deploymentId
      - status
      title: DeploymentResponse
      type: object
    DeploymentStatus:
      example:
        deploymentId: 123e4567-e89b-12d3-a456-426614174000
        progress: 75
        startTime: 2025-04-21T14:30:00Z
        endTime: 2025-04-21T14:35:00Z
        message: Deployment in progress
        results:
          apexTestResults:
          - testRunStatus: Passed
            testRunId: 123e4567-e89b-12d3-a456-426614174000
            message: Test completed successfully
            testRunStackTrace: ""
            testRunDuration: 0.123456789
            testRunEndTime: 2025-04-21T14:35:00Z
            apexClass: MyTestClass
            testEndTime: 2025-04-21T14:35:00Z
            testRunStartTime: 2025-04-21T14:30:00Z
            stackTrace: ""
            testStartTime: 2025-04-21T14:30:00Z
            outcome: Passed
            testDuration: 0.123456789
            testRunMessage: Test run completed successfully
          - testRunStatus: Passed
            testRunId: 123e4567-e89b-12d3-a456-426614174000
            message: Test completed successfully
            testRunStackTrace: ""
            testRunDuration: 0.123456789
            testRunEndTime: 2025-04-21T14:35:00Z
            apexClass: MyTestClass
            testEndTime: 2025-04-21T14:35:00Z
            testRunStartTime: 2025-04-21T14:30:00Z
            stackTrace: ""
            testStartTime: 2025-04-21T14:30:00Z
            outcome: Passed
            testDuration: 0.123456789
            testRunMessage: Test run completed successfully
          componentErrors:
          - componentType: CustomField
            problem: Field creation failed
            fullName: Account.CustomField__c
          - componentType: CustomField
            problem: Field creation failed
            fullName: Account.CustomField__c
          success: true
          componentSuccesses:
          - componentType: CustomField
            problem: Field creation failed
            fullName: Account.CustomField__c
          - componentType: CustomField
            problem: Field creation failed
            fullName: Account.CustomField__c
          completedDate: 2025-04-21T14:35:00Z
          componentFailures:
          - componentType: CustomField
            problem: Field creation failed
            fullName: Account.CustomField__c
          - componentType: CustomField
            problem: Field creation failed
            fullName: Account.CustomField__c
        status: in_progress
      properties:
        deploymentId:
          description: Deployment ID
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: deploymentId
          type: string
        status:
          description: Deployment status
          enum:
          - queued
          - in_progress
          - completed
          - failed
          - cancelled
          example: in_progress
          title: status
          type: string
        progress:
          description: Progress percentage
          example: 75
          maximum: 100
          minimum: 0
          title: progress
          type: integer
        startTime:
          description: Deployment start time
          example: 2025-04-21T14:30:00Z
          format: date-time
          title: startTime
          type: string
        endTime:
          description: Deployment end time
          example: 2025-04-21T14:35:00Z
          format: date-time
          title: endTime
          type: string
        message:
          description: Additional information
          example: Deployment in progress
          title: message
          type: string
        results:
          $ref: '#/components/schemas/DeploymentStatus_results'
      required:
      - deploymentId
      - status
      title: DeploymentStatus
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http